pyproject.toml
[tool.poetry]
name = "api-diplom"
version = "0.1.0"
description = ""
authors = ["KateVarg <katye9470@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
requests = "^2.32.3"
jsonschema = "^4.23.0"
allure-pytest = "^2.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = """
        --alluredir=allure-results
        """